src/main/java/by/prokopovich/time_tracker/service/UserService.java
package by.prokopovich.time_tracker.service;

import by.prokopovich.time_tracker.repository.UserRepository;
import lombok.RequiredArgsConstructor;
import org.springframework.security.core.userdetails.UserDetails;
import org.springframework.security.core.userdetails.UserDetailsService;
import org.springframework.security.core.userdetails.UsernameNotFoundException;
import org.springframework.stereotype.Service;

@Service
@RequiredArgsConstructor
public class UserService {

    private final UserRepository userRepository;

    public UserDetailsService userDetailsService(){
        return new UserDetailsService() {
            @Override
            public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
               return userRepository.findByEmail(username)
                        .orElseThrow(() -> new UsernameNotFoundException(String.format("Пользователь %s не найден", username)));
            }
        };
    }
}
